Set up environment for proxy server
Update README to include instructions for setting up the repos.

Set up server.js with a port that does not conflict with the ports from the other 3 repos.

Set up webpack config.

Set up index.html with scripts and placeholder divs for each of the apps that will be loaded onto the page.